7/23 1552

+++ java/src/main/java/com/techelevator/Candy.java
@@ -1,9 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends Item {
+
     private String type = "Candy";
 
-    public Candy(String slot, String name, double price) {
+    public Candy(String slot, String name, BigDecimal price) {
         super(slot, name, price);
     }
