+++ java/src/main/java/com/techelevator/Chip.java
@@ -1,10 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Chip extends Item {
 
     private String type = "Chip";
 
-    public Chip(String slot, String name, double price) {
+    public Chip(String slot, String name, BigDecimal price) {
         super(slot, name, price);
     }
