+++ java/src/main/java/com/techelevator/Drink.java
@@ -1,10 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Drink extends Item {
 
     private String type = "Drink";
 
-    public Drink(String slot, String name, double price) {
+    public Drink(String slot, String name, BigDecimal price) {
         super(slot, name, price);
     }
