+++ java/src/main/java/com/techelevator/Gum.java
@@ -1,9 +1,11 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Gum extends Item {
     private String type = "Gum";
 
-    public Gum(String slot, String name, double price) {
+    public Gum(String slot, String name, BigDecimal price) {
         super(slot, name, price);
     }
