+++ java/src/main/java/com/techelevator/Item.java
@@ -1,16 +1,17 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
 
 public abstract class Item {
 
     private String slot;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String type;
 
     private int inventory;
 
-    public Item(String slot, String name, Double price) {
+    public Item(String slot, String name, BigDecimal price) {
         this.inventory = 5;
         this.slot = slot;
         this.name = name;
@@ -31,7 +32,7 @@ public abstract class Item {
     public String getName() {
         return name;
     }
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
     public String getType() {
